立ち絵の表示処理を行う場所を変更  #1456
