+++ src/components/accessories/DisplaySettingSampleView.vue
@@ -16,8 +16,8 @@ import type {
   tatieOrderListType,
   dataTextType,
 } from '@/type/data-type'
-import { SelectTatieIndexHigherUpData } from '@/utils/analysisData'
-import { watch, ref } from 'vue'
+import { SelectTatieIndexHigherUpData, createVoiceFileEncodeSetting } from '@/utils/analysisData'
+import { watch, onUnmounted, ref } from 'vue'
 import DisplayMoviePicFile from '@/components/accessories/DisplayMoviePicFile.vue'
 import { DEFAULT_KYARA_TATIE_UUID } from '@/data/data'
 import { MakeClassString } from '@/utils/analysisGeneral'
@@ -43,6 +43,28 @@ const selectTatiePicFile = (): string => {
 // 設定によって立ち絵画像のUUIDを変更する
 const setTatiePicFile = ref<string>(selectTatiePicFile())
 
+////  コンポーネント表示時に、立ち絵画像の表示をも行う
+// 表示するプロファイル情報
+const profile = ref<outSettingType>(createVoiceFileEncodeSetting(props.selectKyara, props.dateList))
+// 設定変更の比較チェックのため、内容を文字列に変換して保存する変数
+const checkConf = ref<string>(JSON.stringify([profile.value?.tatie, tatieSituation], undefined, 2))
+
+// 指定時間ごとに確認し、立ち絵の設定が変わったら表示を変更する
+const onEncodeTatie = setInterval(() => {
+  // 比較のために設定内容をJSON形式に変換
+  const ans = JSON.stringify(
+    [createVoiceFileEncodeSetting(props.selectKyara, props.dateList)?.tatie, tatieSituation],
+    undefined,
+    2,
+  )
+
+  // 比較して前回の内容と異なっていれば立ち絵画像の表示を更新する。
+  if (checkConf.value !== ans) {
+    profile.value = createVoiceFileEncodeSetting(props.selectKyara, props.dateList)
+    checkConf.value = ans
+  }
+}, 2000)
+
 // 選択しているキャラ設定が変更されたら変更する
 watch(
   () => {
@@ -62,6 +84,11 @@ watch(
   },
   { deep: true },
 )
+
+// コンポーネントが表示されなくなったらsetIntervalを停止
+onUnmounted(() => {
+  clearInterval(onEncodeTatie)
+})
 </script>
 
 <template>
@@ -92,15 +119,13 @@ watch(
         待機中
       </button>
     </div>
-    <div
-      class="h-36 w-full border-[1px] border-gray-400"
-      v-if="setTatiePicFile !== DEFAULT_KYARA_TATIE_UUID"
-    >
+    <div class="h-36 w-full border-[1px] border-gray-400" v-if="setTatiePicFile !== DEFAULT_KYARA_TATIE_UUID">
       <DisplayMoviePicFile
         :dateList="dateList"
         :settype="settype"
         :selectKyara="selectKyara"
         :infoData="infoData"
+        :profile="profile"
         :tatieOrderList="tatieOrderList"
         :tatieSituation="tatieSituation"
         imgClass="w-full h-full"
